only scanning functionality in RedhawkGCInterface
