+++ src/coreclr/nativeaot/Runtime/GCMemoryHelpers.cpp
@@ -9,7 +9,6 @@
 #include "gcenv.h"
 #include "PalRedhawkCommon.h"
 #include "CommonMacros.inl"
-
 #include "GCMemoryHelpers.inl"
 
 // This function clears a piece of memory in a GC safe way.
